chore update models and dto
