+++ Singer.API/ClientApp/package.json
@@ -1,6 +1,6 @@
 {
    "name": "Singer",
-   "version": "0.1.0-issue-AB-28-registrationcount.863",
+   "version": "0.1.0-issue-AB-28-registrationcount.864",
    "scripts": {
       "ng": "ng",
       "start": "ng serve",
