+++ Singer.API/ClientApp/src/app/modules/core/models/eventslot.dto.ts
@@ -1,8 +1,3 @@
-export interface EventSlot {
-   id: string;
-   startDateTime: Date;
-   endDateTime: Date;
-}
 export interface EventSlotDTO {
    id: string;
    startDateTime: Date;
