+++ Singer.API/ClientApp/src/app/modules/core/models/singer-event-location.dto.ts
