+++ Singer.API/ClientApp/src/app/modules/core/services/singerevents-api/singerevent-location.proxy.ts
@@ -8,7 +8,7 @@ import {
    UpdateSingerEventLocationDTO,
    CreateSingerEventLocationDTO,
    SingerEventLocationDTO,
-} from '../../models/singer-event-location';
+} from '../../models/singer-event-location.dto';
 
 @Injectable({
    providedIn: 'root',
