+++ Singer.API/ClientApp/src/app/modules/dashboard/components/home/home.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { EventDescription } from 'src/app/modules/core/models/singerevent.model';
 import { SearchEventData } from '../event-search/event-search.component';
-import { SingerEventLocation } from 'src/app/modules/core/models/singer-event-location';
+import { SingerEventLocation } from 'src/app/modules/core/models/singer-event-location.dto';
 import { AuthService } from 'src/app/modules/core/services/auth.service';
 import { SingerEventsService } from 'src/app/modules/core/services/singerevents-api/singerevents.service';
 import { SingerEventLocationService } from 'src/app/modules/core/services/singerevents-api/singerevent-location.service';
